remove borders on comment cells
