+++ Classes/Issues/Comments/IssueCommentBaseCell.swift
@@ -16,20 +16,9 @@ class IssueCommentBaseCell: UICollectionViewCell, UIGestureRecognizerDelegate {
 
     static let collapseCellMinHeight: CGFloat = 30
 
-    enum BorderType {
-        case head
-        case neck
-        case tail
-    }
-
     weak var doubleTapDelegate: IssueCommentDoubleTapDelegate?
-    var border: BorderType = .neck {
-        didSet { setNeedsLayout() }
-    }
     let doubleTapGesture = UITapGestureRecognizer()
 
-    private let borderLayer = CAShapeLayer()
-    private let backgroundLayer = CAShapeLayer()
     private let collapseLayer = CAGradientLayer()
     private let collapseButton = UIButton()
 
@@ -42,18 +31,6 @@ class IssueCommentBaseCell: UICollectionViewCell, UIGestureRecognizerDelegate {
         doubleTapGesture.numberOfTapsRequired = 2
         doubleTapGesture.delegate = self
         addGestureRecognizer(doubleTapGesture)
-
-        // insert above contentView layer
-        borderLayer.strokeColor = UIColor.red.cgColor
-        borderLayer.strokeColor = Styles.Colors.Gray.border.color.cgColor
-        borderLayer.lineWidth = 1 / UIScreen.main.scale
-        borderLayer.fillColor = nil
-        layer.addSublayer(borderLayer)
-
-        // insert as base layer
-        backgroundLayer.strokeColor = nil
-        backgroundLayer.fillColor = UIColor.white.cgColor
-        layer.insertSublayer(backgroundLayer, at: 0)
 
         collapseLayer.isHidden = true
         collapseLayer.colors = [
@@ -84,69 +61,16 @@ class IssueCommentBaseCell: UICollectionViewCell, UIGestureRecognizerDelegate {
         super.layoutSubviews()
         layoutContentViewForSafeAreaInsets()
 
-        let bounds = contentView.frame
-        let inset = borderLayer.lineWidth / 2
-        let pixelSnapBounds = bounds.insetBy(dx: inset, dy: inset)
-        let cornerRadius = Styles.Sizes.cardCornerRadius
-
-        let borderPath = UIBezierPath()
-        let fillPath: UIBezierPath
-
-        switch border {
-        case .head:
-            borderPath.move(to: CGPoint(x: pixelSnapBounds.minX, y: bounds.maxY))
-            borderPath.addLine(to: CGPoint(x: pixelSnapBounds.minX, y: pixelSnapBounds.minY + cornerRadius))
-            borderPath.addQuadCurve(
-                to: CGPoint(x: pixelSnapBounds.minX + cornerRadius, y: pixelSnapBounds.minY),
-                controlPoint: CGPoint(x: pixelSnapBounds.minX, y: pixelSnapBounds.minY)
-            )
-            borderPath.addLine(to: CGPoint(x: pixelSnapBounds.maxX - cornerRadius, y: pixelSnapBounds.minY))
-            borderPath.addQuadCurve(
-                to: CGPoint(x: pixelSnapBounds.maxX, y: pixelSnapBounds.minY + cornerRadius),
-                controlPoint: CGPoint(x: pixelSnapBounds.maxX, y: pixelSnapBounds.minY)
-            )
-            borderPath.addLine(to: CGPoint(x: pixelSnapBounds.maxX, y: bounds.maxY))
-
-            fillPath = borderPath.copy() as! UIBezierPath
-            fillPath.close()
-        case .neck:
-            borderPath.move(to: CGPoint(x: pixelSnapBounds.minX, y: bounds.minY))
-            borderPath.addLine(to: CGPoint(x: pixelSnapBounds.minX, y: bounds.maxY))
-            borderPath.move(to: CGPoint(x: pixelSnapBounds.maxX, y: bounds.minY))
-            borderPath.addLine(to: CGPoint(x: pixelSnapBounds.maxX, y: bounds.maxY))
-
-            fillPath = UIBezierPath(rect: bounds)
-        case .tail:
-            borderPath.move(to: CGPoint(x: pixelSnapBounds.minX, y: bounds.minY))
-            borderPath.addLine(to: CGPoint(x: pixelSnapBounds.minX, y: pixelSnapBounds.maxY - cornerRadius))
-            borderPath.addQuadCurve(
-                to: CGPoint(x: pixelSnapBounds.minX + cornerRadius, y: pixelSnapBounds.maxY),
-                controlPoint: CGPoint(x: pixelSnapBounds.minX, y: pixelSnapBounds.maxY)
-            )
-            borderPath.addLine(to: CGPoint(x: pixelSnapBounds.maxX - cornerRadius, y: pixelSnapBounds.maxY))
-            borderPath.addQuadCurve(
-                to: CGPoint(x: pixelSnapBounds.maxX, y: pixelSnapBounds.maxY - cornerRadius),
-                controlPoint: CGPoint(x: pixelSnapBounds.maxX, y: pixelSnapBounds.maxY)
-            )
-            borderPath.addLine(to: CGPoint(x: pixelSnapBounds.maxX, y: bounds.minY))
-
-            fillPath = borderPath.copy() as! UIBezierPath
-            fillPath.close()
-        }
-        borderLayer.path = borderPath.cgPath
-        backgroundLayer.path = fillPath.cgPath
-
-        borderLayer.frame = self.bounds
-        backgroundLayer.frame = self.bounds
+        let contentBounds = contentView.bounds
 
         if collapseLayer.isHidden == false {
             contentView.layer.addSublayer(collapseLayer)
             contentView.bringSubview(toFront: collapseButton)
 
             let collapseFrame = CGRect(
-                x: bounds.minX,
-                y: bounds.height - IssueCommentBaseCell.collapseCellMinHeight,
-                width: bounds.width,
+                x: contentBounds.minX,
+                y: contentBounds.height - IssueCommentBaseCell.collapseCellMinHeight,
+                width: contentBounds.width,
                 height: IssueCommentBaseCell.collapseCellMinHeight
             )
 
@@ -161,14 +85,6 @@ class IssueCommentBaseCell: UICollectionViewCell, UIGestureRecognizerDelegate {
                 y: collapseFrame.maxY - collapseButton.bounds.height / 2
             )
         }
-    }
-
-    override var backgroundColor: UIColor? {
-        get {
-            guard let color = backgroundLayer.fillColor else { return nil }
-            return UIColor(cgColor: color)
-        }
-        set { backgroundLayer.fillColor = newValue?.cgColor}
     }
 
     override func prepareForReuse() {
