+++ Classes/Issues/Review/IssueReviewEmptyTailCell.swift
@@ -11,15 +11,6 @@ import IGListKit
 
 final class IssueReviewEmptyTailCell: IssueCommentBaseCell, ListBindable {
 
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        border = .tail
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
     // MARK: ListBindable
 
     func bindViewModel(_ viewModel: Any) {}
